Rename otherValidation to nonSchemaValidation
