+++ gbfs-validator/nonSchemaValidation/additional_properties.js
