+++ gbfs-validator/nonSchemaValidation/check_geofencing_zones.js
