+++ gbfs-validator/nonSchemaValidation/check_pricing_plans.js
