+++ gbfs-validator/nonSchemaValidation/check_stations.js
