+++ gbfs-validator/nonSchemaValidation/check_vehicle_types.js
