+++ gbfs-validator/nonSchemaValidation/check_vehicles.js
