+++ gbfs-validator/nonSchemaValidation/index.js
