+++ gbfs-validator/nonSchemaValidation/translations.js
