+++ gbfs-validator/nonSchemaValidation/ttl.js
