+++ gbfs-validator/nonSchemaValidation/utils.js
