+++ gbfs-validator/nonSchemaValidation/vehicle_type_pricing.js
