+++ gbfs-validator/versions/v1.1.js
@@ -1,4 +1,4 @@
-const o = require('../otherValidation')
+const o = require('../nonSchemaValidation')
 
 module.exports = {
   gbfsRequired: false,
